+++ extensions/iceworks-app/src/extension.ts
@@ -156,7 +156,7 @@ export async function activate(context: vscode.ExtensionContext) {
   // init editor title menu
   if (projectType !== 'unknown') {
     vscode.commands.executeCommand('setContext', 'iceworks:showScriptIconInEditorTitleMenu', true);
-    await createEditorMenuAction();
+    await createEditorMenuAction(context, recorder);
   }
 
   // TODO auto start welcome page when the application is new