+++ extensions/iceworks-material-helper/src/services/index.ts
@@ -1,5 +1,6 @@
 import * as common from '@iceworks/common-service';
 import * as block from './block';
+import * as project from '@iceworks/project-service';
 import { page, component, router, menu, material } from '@iceworks/material-engine';
 
-export default { material, common, block, component, router, menu, page };
+export default { material, common, block, component, router, menu, page, project };