+++ packages/material-engine/src/page/templates/template.rax.ts
@@ -1,16 +1,17 @@
 const templateStr = `import { createElement } from 'rax';
+import View from 'rax-view';
 <% for(var i = 0; i < blocks.length; i++) { -%>
 import <%= blocks[i].className %> from '<%= blocks[i].relativePath -%>';
 <% } -%>
 
 export default function () {
   return (
-    <div className="<%= pageName %>-page">
+    <View className="<%= pageName %>-page">
       <% for(var i = 0; i < blocks.length; i++){ -%>
       <% if (blocks[i].description) { %>{/* <%= blocks[i].description -%> */}<% } %>
       <<%= blocks[i].className -%> />
       <% } -%>
-    </div>
+    </View>
   );
 }
 `;